renaming menu-trigger packages to menu
