+++ packages/@react-aria/menu/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@react-aria/menu-trigger",
+  "name": "@react-aria/menu",
   "version": "3.0.0-alpha.2",
   "private": true,
   "description": "Spectrum UI components in React",
