+++ packages/@react-aria/menu/src/useMenuTrigger.ts
