+++ packages/@react-spectrum/menu/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@react-spectrum/menu-trigger",
+  "name": "@react-spectrum/menu",
   "version": "3.0.0-alpha.2",
   "private": true,
   "description": "Spectrum UI components in React",
@@ -34,7 +34,7 @@
     "@react-aria/focus": "^3.0.0-rc.1",
     "@react-aria/i18n": "^3.0.0-rc.1",
     "@react-aria/interactions": "^3.0.0-rc.1",
-    "@react-aria/menu-trigger": "^3.0.0-alpha.2",
+    "@react-aria/menu": "^3.0.0-alpha.2",
     "@react-aria/overlays": "^3.0.0-alpha.2",
     "@react-aria/selection": "^3.0.0-alpha.2",
     "@react-aria/utils": "^3.0.0-rc.1",
